Aramco_AIS_Repo_updated_with_src/Aramco_AIS_Repo/P2_RUL_Predictive_Maintenance/requirements.txt
pandas>=1.5
numpy>=1.21
scikit-learn>=1.2
lifelines>=0.27
xgboost>=1.7
scikit-survival>=0.20
matplotlib>=3.7
mlflow>=2.3
pyyaml>=5.4